\name{cheat}
\alias{cheat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Answers for questions 1, 10, 12, 16 and 17 of Assignment 3.1.
}
\description{
The function provides the answers for questions 1, 10, 12, 16 and 17 of Assignment 3.1.
The assigment was created for the 2022 Programming in Psychological Science course,
which is part of the Research Masters in Psychology degree, at the University of Amsterdam.
}
\usage{
cheat(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
number vector defining the question to be answered
}
}
\details{
cheat uses if statements to execute the code for the respective answers.
For challenging questions the function also prints the code of the answer.
For x other than 1, 10, 12, 16 and 17, it informs the user of which questions are included.
}
\value{
The output of the answer code.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Xanthippi Alexi Vassiliou
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## Give the answer and code for question 17
cheat(17)

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
