\name{remind_me}
\alias{remind_me}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Public Holidays in the Netherlands for the Year 2022
}
\description{
This function prints a list with the dates of all official public holidays in the Netherlands in 2022.
The information was found in the government website:
https://www.government.nl/topics/working-hours/question-and-answer/public-holidays-in-the-netherlands
}
\usage{
remind_me()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  This function takes no arguments
}
\details{
remind_me uses cat (concatenate and print) and paste to print a list of all official public holidays in the Netherlands for the year 2022.
}
\value{
Outputs a character vector of all official public holidays in the Netherlands for the year 2022.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Xanthippi Alexi Vassiliou
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## print public holidays
remind_me()

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
